modules/location-db-handler/requirements.txt
grpcio==1.42.0
grpcio-tools==1.42.0
kafka-python==2.0.2
protobuf==3.19.1
six==1.16.0
